bump yo-fs to 3.0.0
