+++ app.js
@@ -10,9 +10,9 @@ var drive = hyperdrive(db)
 var speedometer = require('speedometer')
 var prettyBytes = require('pretty-bytes')
 var path = require('path')
-var explorer = require('./')
 var intro = require('intro.js')
 var yo = require('yo-yo')
+var explorer = require('./')
 
 var $hyperdrive = document.querySelector('#hyperdrive-ui')
 var $shareLink = document.getElementById('share-link')
@@ -87,29 +87,13 @@ function main (key) {
     window.location = '#' + archive.key.toString('hex')
     updateShareLink()
 
-    var entries = []
-    var widget
-
-    function update () {
-      var fresh = explorer(archive, entries, function (ev, entry) {
-        if (entry.type === 'directory') {
-          cwd = entry.name
-        }
-      })
-      if (widget) widget = yo.update(widget, fresh)
-      else widget = fresh
+    function onclick (ev, entry) {
+      if (entry.type === 'directory') {
+        cwd = entry.name
+      }
     }
-
-    update()
-    $hyperdrive.appendChild(widget)
-
-    var stream = archive.list({live: true})
-    stream.on('data', function (entry) {
-      if (archive.owner) help.innerHTML = 'drag and drop files'
-      else help.innerHTML = ''
-      entries.push(entry)
-      update()
-    })
+    var tree = explorer(archive, onclick)
+    $hyperdrive.appendChild(tree)
   })
 }
